+++ Backend/src/main/kotlin/org/pin/backend/service/PagoService.kt
@@ -1,8 +1,10 @@
 package org.pin.backend.service
 import org.pin.backend.repository.PagoRepository
 import org.springframework.stereotype.Service
+import org.springframework.transaction.annotation.Transactional
 
 @Service
+@Transactional(readOnly = true)
 class PagoService(
     private val repo: PagoRepository,
 ) {